create extractor and integration testing
