+++ waimak-rdbm-ingestion/src/main/scala/com/coxautodata/waimak/rdbm/ingestion/SQLServerViewExtractor.scala
@@ -17,7 +17,7 @@ import scala.util.{Failure, Success, Try}
 class SQLServerViewExtractor(override val sparkSession: SparkSession
                              , sqlServerConnectionDetails: SQLServerConnectionDetails
                              , extraConnectionProperties: Properties = new Properties()
-                             , override val transformTableNameForRead: String => String = identity) extends SQLServerExtractor(sqlServerConnectionDetails, extraConnectionProperties) {
+                             , override val transformTableNameForRead: String => String = identity) extends SQLServerBaseExtractor(sqlServerConnectionDetails, extraConnectionProperties) {
 
   override def getTableMetadata(dbSchemaName: String
                                 , tableName: String
