+++ waimak-rdbm-ingestion/src/test/scala/com/coxautodata/waimak/rdbm/ingestion/SQLServerExtractorIntegrationTest.scala
@@ -15,7 +15,7 @@ import org.scalatest.BeforeAndAfterAll
 import scala.util.Success
 
 /**
-  * Created by Vicky Avison on 19/04/18.
+  * Created by Ian Baynham on 19/04/18.
   */
 class SQLServerExtractorIntegrationTest extends SparkAndTmpDirSpec with BeforeAndAfterAll {
 
@@ -34,54 +34,29 @@ class SQLServerExtractorIntegrationTest extends SparkAndTmpDirSpec with BeforeAn
   }
 
   def setupTables(): Unit = {
-    val testTemporalTableCreate =
-      """CREATE TABLE TestTemporal
+    
+    val testTableCreate =
+      """CREATE TABLE testtable
         |(
-        |     TestTemporalID int NOT NULL PRIMARY KEY CLUSTERED
-        |   , SysStartTime datetime2 GENERATED ALWAYS AS ROW START HIDDEN NOT NULL
-        |   , SysEndTime datetime2 GENERATED ALWAYS AS ROW END HIDDEN NOT NULL
-        |   , TestTemporalValue varchar(50) NOT NULL
-        |   , PERIOD FOR SYSTEM_TIME (SysStartTime, SysEndTime)
-        |)
-        |WITH
-        |   (
-        |      SYSTEM_VERSIONING = ON (HISTORY_TABLE = dbo.TestTemporalHistory)
-        |   )
-        |;
-        |        insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (1, 'Value1');
-        |        insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (2, 'Value2');
-        |        insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (3, 'Value3');
-        |        insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (4, 'Value4');
-        |        insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (5, 'Value5');
-        |        insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (6, 'Value6');
-        |        insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (7, 'Value7');
-        |        update TestTemporal set TestTemporalValue = 'New Value 1' where TestTemporalID = 1""".stripMargin
-
-    val testNonTemporalTableCreate =
-      """CREATE TABLE TestNonTemporal
-        |(
-        |     TestNonTemporalID1 int NOT NULL
-        |   , TestNonTemporalID2 int NOT NULL
-        |   , TestNonTemporalValue varchar(50) NOT NULL
-        |   PRIMARY KEY CLUSTERED(TestNonTemporalID1, TestNonTemporalID2)
+        |     testtableID1 int NOT NULL
+        |   , testtableID2 int NOT NULL
+        |   , testtableValue varchar(50) NOT NULL
+        |   PRIMARY KEY CLUSTERED(testtableID1, testtableID2)
         |)
         |;
         |
-        |   insert into TestNonTemporal (TestNonTemporalID1, TestNonTemporalID2, TestNonTemporalValue) VALUES (1, 1, 'V1');
-        |   insert into TestNonTemporal (TestNonTemporalID1, TestNonTemporalID2, TestNonTemporalValue) VALUES (2, 1, 'V2');
-        |   insert into TestNonTemporal (TestNonTemporalID1, TestNonTemporalID2, TestNonTemporalValue) VALUES (2, 2, 'V3');
-        |   insert into TestNonTemporal (TestNonTemporalID1, TestNonTemporalID2, TestNonTemporalValue) VALUES (4, 3, 'V4');
-        |   insert into TestNonTemporal (TestNonTemporalID1, TestNonTemporalID2, TestNonTemporalValue) VALUES (5, 3, 'V5');""".stripMargin
+        |   insert into testtable (testtableID1, testtableID2, testtableValue) VALUES (1, 1, 'V1');
+        |   insert into testtable (testtableID1, testtableID2, testtableValue) VALUES (2, 1, 'V2');
+        |   insert into testtable (testtableID1, testtableID2, testtableValue) VALUES (2, 2, 'V3');
+        |   insert into testtable (testtableID1, testtableID2, testtableValue) VALUES (4, 3, 'V4');
+        |   insert into testtable (testtableID1, testtableID2, testtableValue) VALUES (5, 3, 'V5');""".stripMargin
 
-    executeSQl(Seq(testTemporalTableCreate, testNonTemporalTableCreate))
+    executeSQl(Seq(testTableCreate))
   }
 
   def cleanupTables(): Unit = {
     executeSQl(Seq(
-      "alter table TestTemporal set (SYSTEM_VERSIONING = OFF)"
-      , "drop table if exists TestTemporal"
-      , "drop table if exists TestTemporalHistory"
-      , "drop table if exists TestNonTemporal"
+       "drop table if exists testtable"
     ))
   }
 
@@ -94,36 +69,19 @@ class SQLServerExtractorIntegrationTest extends SparkAndTmpDirSpec with BeforeAn
   }
 
   describe("getTableMetadata") {
-    it("should read the metadata for a temporal table") {
-      val sqlServerExtractor = new SQLServerTemporalExtractor(sparkSession, sqlServerConnectionDetails)
 
-      sqlServerExtractor.allTableMetadata("dbo.testtemporal") should be(
-        SQLServerTemporalTableMetadata("dbo", "testtemporal", Some("dbo"), Some("testtemporalhistory"), Some("sysstarttime"), Some("sysendtime"), "testtemporalid"))
+    it("should read the metadata for a test table") {
+      val sqlServerExtractor = new SQLServerExtractor(sparkSession, sqlServerConnectionDetails)
 
-      sqlServerExtractor.getTableMetadata("dbo", "testtemporal") should be(
-        Success(AuditTableInfo("testtemporal", Seq("testtemporalid"), Map(
-          "schemaName" -> "dbo"
-          , "tableName" -> "testtemporal"
-          , "primaryKeys" -> "testtemporalid"
-          , "historyTableSchema" -> "dbo"
-          , "historyTableName" -> "testtemporalhistory"
-          , "startColName" -> "sysstarttime"
-          , "endColName" -> "sysendtime")))
-      )
-    }
-
-    it("should read the metadata for a non-temporal table") {
-      val sqlServerExtractor = new SQLServerTemporalExtractor(sparkSession, sqlServerConnectionDetails)
-
-      sqlServerExtractor.allTableMetadata("dbo.testnontemporal") should be(
-        SQLServerTemporalTableMetadata("dbo", "testnontemporal", None, None, None, None, "testnontemporalid1;testnontemporalid2")
+      sqlServerExtractor.allTablePKs("dbo.testtable") should be(
+        sqlServerExtractor.SQLServerTableMetadata("dbo", "testtable", "testtableid1;testtableid2")
       )
 
-      sqlServerExtractor.getTableMetadata("dbo", "testnontemporal") should be(
-        Success(AuditTableInfo("testnontemporal", Seq("testnontemporalid1", "testnontemporalid2"), Map(
+      sqlServerExtractor.getTableMetadata("dbo", "testtable") should be(
+        Success(AuditTableInfo("testtable", Seq("testtableid1", "testtableid2"), Map(
           "schemaName" -> "dbo"
-          , "tableName" -> "testnontemporal"
-          , "primaryKeys" -> "testnontemporalid1;testnontemporalid2")))
+          , "tableName" -> "testtable"
+          , "primaryKeys" -> "testtableid1;testtableid2")))
       )
     }
   }
@@ -133,47 +91,20 @@ class SQLServerExtractorIntegrationTest extends SparkAndTmpDirSpec with BeforeAn
     it("should extract from the db to the storage layer") {
       val spark = sparkSession
       import spark.implicits._
-      val sqlServerExtractor = new SQLServerTemporalExtractor(sparkSession, sqlServerConnectionDetails)
+      val sqlServerExtractor = new SQLServerExtractor(sparkSession, sqlServerConnectionDetails)
       val flow = Waimak.sparkFlow(sparkSession)
       val executor = Waimak.sparkExecutor()
-      val tableConfig = Map("testtemporal" -> RDBMExtractionTableConfig("testtemporal")
-        , "testnontemporal" -> RDBMExtractionTableConfig("testnontemporal"))
-      val writeFlow = flow.extractToStorageFromRDBM(sqlServerExtractor, "dbo", s"$testingBaseDir/output", tableConfig, insertDateTime)("testtemporal", "testnontemporal")
+      val tableConfig = Map("testtable" -> RDBMExtractionTableConfig("testtable"))
+      val writeFlow = flow.extractToStorageFromRDBM(sqlServerExtractor, "dbo", s"$testingBaseDir/output", tableConfig, insertDateTime)("testtable")
 
       val res1 = executor.execute(writeFlow)
-      res1._2.inputs.get[Dataset[_]]("testtemporal").sort("source_type", "testtemporalid")
-        .as[TestTemporal].collect() should be(Seq(
-        TestTemporal(1, "New Value 1", 0)
-        , TestTemporal(2, "Value2", 0)
-        , TestTemporal(3, "Value3", 0)
-        , TestTemporal(4, "Value4", 0)
-        , TestTemporal(5, "Value5", 0)
-        , TestTemporal(6, "Value6", 0)
-        , TestTemporal(7, "Value7", 0)
-        , TestTemporal(1, "Value1", 1)
-      ))
-
-      res1._2.inputs.get[Dataset[_]]("testnontemporal").sort("testnontemporalid1", "testnontemporalid2")
-        .as[TestNonTemporal].collect() should be(Seq(
-        TestNonTemporal(1, 1, "V1")
-        , TestNonTemporal(2, 1, "V2")
-        , TestNonTemporal(2, 2, "V3")
-        , TestNonTemporal(4, 3, "V4")
-        , TestNonTemporal(5, 3, "V5")
-      ))
-
-      val readFlow = flow.loadFromStorage(s"$testingBaseDir/output")("testtemporal")
-      val res2 = executor.execute(readFlow)
-      res2._2.inputs.get[Dataset[_]]("testtemporal").sort("source_type", "testtemporalid")
-        .as[TestTemporal].collect() should be(Seq(
-        TestTemporal(1, "New Value 1", 0)
-        , TestTemporal(2, "Value2", 0)
-        , TestTemporal(3, "Value3", 0)
-        , TestTemporal(4, "Value4", 0)
-        , TestTemporal(5, "Value5", 0)
-        , TestTemporal(6, "Value6", 0)
-        , TestTemporal(7, "Value7", 0)
-        , TestTemporal(1, "Value1", 1)
+      res1._2.inputs.get[Dataset[_]]("testtable").sort("testtableid1", "testtableid2")
+        .as[TestTable].collect() should be(Seq(
+        TestTable(1, 1, "V1")
+        , TestTable(2, 1, "V2")
+        , TestTable(2, 2, "V3")
+        , TestTable(4, 3, "V4")
+        , TestTable(5, 3, "V5")
       ))
     }
 
@@ -181,138 +112,37 @@ class SQLServerExtractorIntegrationTest extends SparkAndTmpDirSpec with BeforeAn
       val spark = sparkSession
       import spark.implicits._
 
-      val sqlServerExtractor = new SQLServerTemporalExtractor(sparkSession, sqlServerConnectionDetails)
+      val sqlServerExtractor = new SQLServerExtractor(sparkSession, sqlServerConnectionDetails)
       val flow = Waimak.sparkFlow(sparkSession)
       val executor = Waimak.sparkExecutor()
-      val tableConfig: Map[String, RDBMExtractionTableConfig] = Map("testtemporal" -> RDBMExtractionTableConfig("testtemporal", maxRowsPerPartition = Some(2))
-        , "testnontemporal" -> RDBMExtractionTableConfig("testnontemporal", maxRowsPerPartition = Some(3)))
+      val tableConfig: Map[String, RDBMExtractionTableConfig] = Map("testtable" -> RDBMExtractionTableConfig("testtable", maxRowsPerPartition = Some(3)))
 
       val writeFlow = flow.extractToStorageFromRDBM(sqlServerExtractor
         , "dbo"
         , s"$testingBaseDir/output"
         , tableConfig
-        , insertDateTime)("testtemporal", "testnontemporal")
+        , insertDateTime)("testtable")
 
       val res = executor.execute(writeFlow)
 
-      val testTemporal = res._2.inputs.get[Dataset[_]]("testtemporal")
+      val testTable = res._2.inputs.get[Dataset[_]]("testtable")
 
-      val testNonTemporal = res._2.inputs.get[Dataset[_]]("testnontemporal")
-
-      testNonTemporal.sort("testnontemporalid1", "testnontemporalid2")
-        .as[TestNonTemporal].collect() should be(Seq(
-        TestNonTemporal(1, 1, "V1")
-        , TestNonTemporal(2, 1, "V2")
-        , TestNonTemporal(2, 2, "V3")
-        , TestNonTemporal(4, 3, "V4")
-        , TestNonTemporal(5, 3, "V5")
+      testTable.sort("testtableid1", "testtableid2")
+        .as[TestTable].collect() should be(Seq(
+        TestTable(1, 1, "V1")
+        , TestTable(2, 1, "V2")
+        , TestTable(2, 2, "V3")
+        , TestTable(4, 3, "V4")
+        , TestTable(5, 3, "V5")
       ))
 
       //Should create two partitions (max rows per partition is 3)
-      testNonTemporal.rdd.getNumPartitions should be(2)
+      testTable.rdd.getNumPartitions should be(2)
 
-      testTemporal.sort("source_type", "testtemporalid")
-        .as[TestTemporal].collect() should be(Seq(
-        TestTemporal(1, "New Value 1", 0)
-        , TestTemporal(2, "Value2", 0)
-        , TestTemporal(3, "Value3", 0)
-        , TestTemporal(4, "Value4", 0)
-        , TestTemporal(5, "Value5", 0)
-        , TestTemporal(6, "Value6", 0)
-        , TestTemporal(7, "Value7", 0)
-        , TestTemporal(1, "Value1", 1)
-      ))
-
-      //Create one partition for the history table and another 4 for the main (split points 2, 4, 6)
-      testTemporal.rdd.getNumPartitions should be(5)
     }
   }
 
-  it("should handle start/stop delta logic for the temporal tables") {
-    val spark = sparkSession
-    import spark.implicits._
-
-    val sqlServerExtractor = new SQLServerTemporalExtractor(sparkSession, sqlServerConnectionDetails)
-    val flow = Waimak.sparkFlow(sparkSession)
-    val executor = Waimak.sparkExecutor()
-
-    val tableConfig: Map[String, RDBMExtractionTableConfig] = Map("testtemporal" -> RDBMExtractionTableConfig("testtemporal", maxRowsPerPartition = Some(2)))
-
-
-    val writeFlow = flow.extractToStorageFromRDBM(sqlServerExtractor
-      , "dbo"
-      , s"$testingBaseDir/output"
-      , tableConfig
-      , insertDateTime)("testtemporal")
-
-    executor.execute(writeFlow)
-
-
-    val deletes = "delete from TestTemporal where TestTemporalId = 2;"
-    val inserts =
-      s"""
-         |insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (8, 'Value8');
-         |insert into TestTemporal (TestTemporalID, TestTemporalValue) VALUES (9, 'Value9');
-         |""".stripMargin
-    val updates =
-      s"""
-         |update TestTemporal set TestTemporalValue = 'New Value 5' where TestTemporalID = 5;
-         |update TestTemporal set TestTemporalValue = 'New Value 6' where TestTemporalID = 6;
-         |update TestTemporal set TestTemporalValue = 'New Value 7' where TestTemporalID = 7;
-       """.stripMargin
-
-    executeSQl(Seq(updates, inserts, deletes))
-
-    val deltaWriteFlow = flow.extractToStorageFromRDBM(sqlServerExtractor
-      , "dbo"
-      , s"$testingBaseDir/output"
-      , tableConfig
-      , insertDateTime)("testtemporal")
-
-    val res = executor.execute(deltaWriteFlow)
-
-    val testTemporal = res._2.inputs.get[Dataset[_]]("testtemporal")
-
-    testTemporal.sort("source_type", "testtemporalid")
-      .as[TestTemporal].collect()
-      //For some reason, sometimes (not consistently) > seems to act like >= on these datetime2 fields so we need to filter
-      //out the records which could mess up our test
-      .filterNot(_.testtemporalid == 1) should be(Seq(
-      TestTemporal(5, "New Value 5", 0)
-      , TestTemporal(6, "New Value 6", 0)
-      , TestTemporal(7, "New Value 7", 0)
-      , TestTemporal(8, "Value8", 0)
-      , TestTemporal(9, "Value9", 0)
-      , TestTemporal(2, "Value2", 1)
-      , TestTemporal(5, "Value5", 1)
-      , TestTemporal(6, "Value6", 1)
-      , TestTemporal(7, "Value7", 1)
-    ))
-
-    val maxTS = Timestamp.valueOf(testTemporal.agg(max($"system_timestamp_of_extraction")).as[String].collect().head)
-
-    val snapshotReadFlow =
-      flow.snapshotTemporalTablesFromStorage(s"$testingBaseDir/output", maxTS)("testtemporal")
-
-    val snapshotRes = executor.execute(snapshotReadFlow)
-
-    val testTemporalSnapshot = snapshotRes._2.inputs.get[Dataset[_]]("testtemporal")
-
-    testTemporalSnapshot.sort("testtemporalid")
-      .as[TestTemporal].collect() should be(Seq(
-      TestTemporal(1, "New Value 1", 0)
-      , TestTemporal(3, "Value3", 0)
-      , TestTemporal(4, "Value4", 0)
-      , TestTemporal(5, "New Value 5", 0)
-      , TestTemporal(6, "New Value 6", 0)
-      , TestTemporal(7, "New Value 7", 0)
-      , TestTemporal(8, "Value8", 0)
-      , TestTemporal(9, "Value9", 0)
-    ))
-
-  }
 }
 
-case class TestTemporal(testtemporalid: Int, testtemporalvalue: String, source_type: Int)
+case class TestTable(testtableid1: Int, testtableid2: Int, testtablevalue: String)
 
-case class TestNonTemporal(testnontemporalid1: Int, testnontemporalid2: Int, testnontemporalvalue: String)
